requirements.txt
cachetools==5.3.2
certifi==2024.2.2
chaostoolkit-google-cloud-platform==0.14.1
chaostoolkit-kubernetes==0.34.1
chaostoolkit-lib==1.41.0
charset-normalizer==3.3.2
dateparser==1.2.0
filelock==3.13.1
fsspec==2024.2.0
generativeai==0.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.4.0
google-cloud-artifact-registry==1.11.0
google-cloud-audit-log==0.2.5
google-cloud-compute==1.16.0
google-cloud-container==2.39.0
google-cloud-containeranalysis==2.14.0
google-cloud-core==2.4.1
google-cloud-logging==3.9.0
google-cloud-monitoring==2.19.0
google-cloud-run==0.10.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-generativeai==0.3.2
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grafeas==1.9.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
kubernetes==29.0.0
logzero==1.7.0
oauthlib==3.2.2
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.0
websocket-client==1.7.0
zipp==3.17.0
